switch source to ESM, migrate off Jest to UVU, drop Node 16
