+++ .github/workflows/main.yml
@@ -8,7 +8,7 @@ jobs:
       pull-requests: write
     steps:
       - name: '🚚 Checkout'
-        uses: actions/checkout@v3
+        uses: actions/checkout@v4
       - name: '🔏 Yarn Lock Changes'
         uses: ./
         with:
