+++ .github/workflows/tests.yml
@@ -6,7 +6,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        node: [16, 18, 20]
+        node: [18, 20, 21]
     name: Testing on Node ${{ matrix.node }}
     steps:
     - name: '🚚 Checkout'
