--- jest.config.js
@@ -1,3 +1,0 @@
-module.exports = {
-  transformIgnorePatterns: ['<rootDir>/node_modules/*']
-};
