+++ package.json
@@ -9,7 +9,7 @@
     "build": "ncc build ./src/action.js -m --target es2021",
     "lint": "eslint .",
     "optimize": "svgo -f ./assets",
-    "test": "jest"
+    "test": "uvu tests -i .cjs -i testUtils"
   },
   "dependencies": {
     "@actions/core": "^1.10.1",
@@ -19,17 +19,14 @@
     "semver": "^7.5.4"
   },
   "devDependencies": {
-    "@babel/core": "^7.23.7",
-    "@babel/preset-env": "^7.23.8",
     "@vercel/ncc": "^0.38.1",
     "@yarnpkg/lockfile": "^1.1.0",
-    "babel-jest": "^29.7.0",
     "eslint": "^8.56.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-jest": "^27.6.3",
     "eslint-plugin-prettier": "^5.1.3",
-    "jest": "^29.7.0",
     "prettier": "^3.2.4",
-    "svgo": "^3.2.0"
+    "svgo": "^3.2.0",
+    "uvu": "^0.5.6"
   }
 }
